change data type of labels

+++ include/caffe/util/io.hpp
@@ -182,7 +182,7 @@ void CVMatToDatum(const cv::Mat& cv_img, Datum* datum);
 
 // for multilabel
 void ReadImagesList(const string& source,
-    std::vector<std::pair<std::string, std::vector<int> > >* images_vec);
+    std::vector<std::pair<std::string, std::vector<float> > >* images_vec);
 
 #endif  // USE_OPENCV
